more1234

+++ everybit/bitarray.c
@@ -205,17 +205,17 @@ void bitarray_rotate(bitarray_t *const bitarray,
   if (d == 0) {
     return;
   }
-  int ind_at_0 = d;
-  int partition = bit_offset + ind_at_0 - 1;
+  size_t ind_at_0 = d;
+  size_t partition = bit_offset + ind_at_0 - 1;
   size_t end =bit_offset + bit_length - 1;
   reverse(bitarray_rotate,bit_offset,partition);
   reverse(bitarray_rotate,partition+1,end);
   reverse(bitarray_rotate,bit_offset,end);
 }
 
-void reverse(bitarray_t *const bitarray, int start, int stop){
-	int j = stop;
-	int i = start;
+void reverse(bitarray_t *const bitarray, size_t start, size_t stop){
+	size_t j = stop;
+	size_t i = start;
 	while( i < j){
 		bool temp = bitarray_get(bitarray,i);
 		bitarray_set(bitarray,i,bitarray_get(bitarray,j));
